Fix rounding for values >1 when decrementing
When trying to set brightness value >1 xrandr will round up the value to 1
decimal place. When using decrement step <0.1, it would cause the value to be
locked at previous value due to this rounding.
In cases when value should be decremented and resulting value would be >1, it is
now rounded down instead (eg. 1.1-0.5=1.0).